Jenkinsfile: parses nodes and tests from YAML
This version parses the nodes and tests to run from the YAML.
It does the 'node()' assignments, but only echos the tests right now.

Problem: we cannot parallel inside a function it seems, so next version
will probably unroll all of that back to the main pipeline and see
if that works?

Signed-off-by: Graham Whaley <graham.whaley@intel.com>